feat: removed duplicate functions and moved to util
