+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/likes/view/LMFeedLikesFragment.kt
@@ -1,12 +1,8 @@
 package com.likeminds.feed.android.core.likes.view
 
-import android.annotation.SuppressLint
-import android.os.Build
 import android.os.Bundle
 import android.view.*
 import androidx.core.content.ContextCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
@@ -87,7 +83,10 @@ open class LMFeedLikesFragment : Fragment(), LMFeedLikesAdapterListener {
     protected open fun customizeLikesFragmentHeaderView(headerViewLikes: LMFeedHeaderView) {
         headerViewLikes.apply {
             setStyle(LMFeedStyleTransformer.likesFragmentViewStyle.headerViewStyle)
-            setStatusBarColor(LMFeedStyleTransformer.likesFragmentViewStyle.headerViewStyle.backgroundColor)
+            LMFeedViewUtils.setStatusBarColor(
+                requireActivity(),
+                LMFeedStyleTransformer.likesFragmentViewStyle.headerViewStyle.backgroundColor
+            )
 
             setTitleText(
                 getString(
@@ -96,29 +95,6 @@ open class LMFeedLikesFragment : Fragment(), LMFeedLikesAdapterListener {
                         .pluralizeOrCapitalize(LMFeedWordAction.FIRST_LETTER_CAPITAL_PLURAL)
                 )
             )
-        }
-    }
-
-    @SuppressLint("InlinedApi")
-    @Suppress("Deprecation")
-    private fun setStatusBarColor(backgroundColor: Int) {
-        val window = requireActivity().window
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
-            windowInsetsController.isAppearanceLightStatusBars = true
-            window.statusBarColor = ContextCompat.getColor(requireContext(), backgroundColor)
-        } else {
-            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
-            insetsController.isAppearanceLightStatusBars = true
-
-            window.decorView.setBackgroundColor(
-                ContextCompat.getColor(
-                    requireContext(),
-                    backgroundColor
-                )
-            )
-            window.insetsController?.hide(WindowInsets.Type.statusBars())
-            window.insetsController?.show(WindowInsets.Type.statusBars())
         }
     }
