+++ likeminds-feed-android-core/src/main/java/com/likeminds/feed/android/core/poll/result/view/LMFeedPollResultsFragment.kt
@@ -1,15 +1,8 @@
 package com.likeminds.feed.android.core.poll.result.view
 
-import android.annotation.SuppressLint
-import android.os.Build
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.view.WindowInsets
+import android.view.*
 import androidx.core.content.ContextCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.Fragment
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
@@ -22,10 +15,7 @@ import com.likeminds.feed.android.core.poll.result.view.LMFeedPollResultsActivit
 import com.likeminds.feed.android.core.ui.base.styles.setStyle
 import com.likeminds.feed.android.core.ui.base.views.LMFeedTextView
 import com.likeminds.feed.android.core.ui.widgets.headerview.view.LMFeedHeaderView
-import com.likeminds.feed.android.core.utils.LMFeedExtrasUtil
-import com.likeminds.feed.android.core.utils.LMFeedStyleTransformer
-import com.likeminds.feed.android.core.utils.LMFeedViewUtils
-import com.likeminds.feed.android.core.utils.emptyExtrasException
+import com.likeminds.feed.android.core.utils.*
 
 open class LMFeedPollResultsFragment : Fragment() {
 
@@ -93,7 +83,7 @@ open class LMFeedPollResultsFragment : Fragment() {
             val headerViewStyle =
                 LMFeedStyleTransformer.pollResultsFragmentViewStyle.headerViewStyle
             setStyle(headerViewStyle)
-            setStatusBarColor(headerViewStyle.backgroundColor)
+            LMFeedViewUtils.setStatusBarColor(requireActivity(), headerViewStyle.backgroundColor)
             binding.tabLayoutPollResults.setBackgroundColor(
                 ContextCompat.getColor(
                     requireContext(),
@@ -102,29 +92,6 @@ open class LMFeedPollResultsFragment : Fragment() {
             )
 
             setTitleText(getString(R.string.lm_feed_poll_results))
-        }
-    }
-
-    @SuppressLint("InlinedApi")
-    @Suppress("Deprecation")
-    private fun setStatusBarColor(backgroundColor: Int) {
-        val window = requireActivity().window
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-            val windowInsetsController = WindowCompat.getInsetsController(window, window.decorView)
-            windowInsetsController.isAppearanceLightStatusBars = true
-            window.statusBarColor = ContextCompat.getColor(requireContext(), backgroundColor)
-        } else {
-            val insetsController = WindowInsetsControllerCompat(window, window.decorView)
-            insetsController.isAppearanceLightStatusBars = true
-
-            window.decorView.setBackgroundColor(
-                ContextCompat.getColor(
-                    requireContext(),
-                    backgroundColor
-                )
-            )
-            window.insetsController?.hide(WindowInsets.Type.statusBars())
-            window.insetsController?.show(WindowInsets.Type.statusBars())
         }
     }
